+++ local.sh
@@ -1,8 +1,9 @@
-export APP_REDIS='redis://localhost:6379/0' \
-       APP_PORT=5000 \
-       STELLAR_HORIZON_URL='https://horizon-kik.kininfrastructure.com' \
-       STELLAR_NETWORK='private testnet' \
-       STELLAR_KIN_ISSUER_ADDRESS='GBQ3DQOA7NF52FVV7ES3CR3ZMHUEY4LTHDAQKDTO6S546JCLFPEQGCPK' \
-       STELLAR_KIN_TOKEN_NAME='KIN' \
-       KIN_FAUCET='http://159.65.84.173:5000' \
-       XLM_FAUCET='http://friendbot-kik.kininfrastructure.com'
+export \
+    APP_REDIS='redis://localhost:6379/0' \
+    APP_PORT=5000 \
+    STELLAR_HORIZON_URL='https://horizon-kik.kininfrastructure.com' \
+    STELLAR_NETWORK='private testnet' \
+    STELLAR_KIN_ISSUER_ADDRESS='GBQ3DQOA7NF52FVV7ES3CR3ZMHUEY4LTHDAQKDTO6S546JCLFPEQGCPK' \
+    STELLAR_KIN_TOKEN_NAME='KIN' \
+    KIN_FAUCET='http://159.65.84.173:5000' \
+    XLM_FAUCET='http://friendbot-kik.kininfrastructure.com'
