+++ payment/utils.py
@@ -1,11 +1,10 @@
 import contextlib
 import time
 from functools import wraps
-from .redis_conn import redis_conn
 
 
 @contextlib.contextmanager
-def lock(key):
+def lock(redis_conn, key):
     _lock = redis_conn.lock('__lock:{}'.format(key), blocking_timeout=10)
     _lock.acquire()
     yield
@@ -26,3 +25,15 @@ def retry(times, delay=0.3):
                         time.sleep(delay)
         return wrapper
     return decorator
+
+
+def get_network_name(stellar_network):
+    """hack: monkeypatch stellar_base to support private network."""
+    if stellar_network in ['PUBLIC', 'TESTNET']:
+        return stellar_network
+    else:
+        import stellar_base
+        PRIVATE = 'PRIVATE'
+        # register the private network with the given passphrase
+        stellar_base.network.NETWORKS[PRIVATE] = stellar_network
+        return PRIVATE
